Separated logic from addresssearch component to a container. Implement more tests.
